move WorkingWithNodeWsTest and WorkingWithJavaMySqlStackTest to stack package
