+++ selenium/che-selenium-test/src/test/resources/suites/CheSuite.xml
@@ -274,8 +274,8 @@
           <class name="org.eclipse.che.selenium.workspaces.ProjectStateAfterRefreshTest"/>
           <class name="org.eclipse.che.selenium.workspaces.ProjectStateAfterRenameWorkspaceTest"/>
           <class name="org.eclipse.che.selenium.workspaces.ProjectStateAfterWorkspaceRestartTest"/>
-          <class name="org.eclipse.che.selenium.workspaces.WorkingWithJavaMySqlStackTest"/>
-          <class name="org.eclipse.che.selenium.workspaces.WorkingWithNodeWsTest"/>
+          <class name="org.eclipse.che.selenium.stack.WorkingWithJavaMySqlStackTest"/>
+          <class name="org.eclipse.che.selenium.stack.WorkingWithNodeWsTest"/>
           <class name="org.eclipse.che.selenium.workspaces.WorkspaceFromOfficialUbuntuImageStartTest"/>
           <class name="org.eclipse.che.selenium.workspaces.CheckStoppingWsByTimeoutTest"/>
           <class name="org.eclipse.che.selenium.dashboard.AccountTest"/>
